src/ingesta/requirements.txt
yfinance>=0.2.36
google-cloud-bigquery>=3.17.0
pandas>=2.0.0
pyarrow>=14.0.0
db-dtypes>=1.2.0
